--- apps/web/src/components/avatar/index.ts
@@ -1,1 +1,0 @@
-export * from "./user";
