--- apps/web/src/components/popover/userMenu/index.ts
@@ -1,1 +1,0 @@
-export * from "./userMenu";
