+++ apps/web/src/components/userPanel/index.ts
@@ -1,0 +1,1 @@
+export * from "./userPanel";
